Move export components to their own directory
